Split disk-backed ART code
